+++ app/src/main/java/com/pavlov/nearWarSecrets/ui/theme/uiComponents/BackgroundImageBase.kt
@@ -1,0 +1,28 @@
+package com.pavlov.nearWarSecrets.ui.theme.uiComponents
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.pavlov.nearWarSecrets.R
+
+@Composable
+fun BackgroundImageBase() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp)
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.cat_background2),
+            contentDescription = null,
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+    }
+}
